Redis 2.6 for Windows

+++ src/intset.c
@@ -85,10 +85,10 @@ static void _intsetSet(intset *is, int pos, int64_t value) {
         ((int64_t*)is->contents)[pos] = value;
         memrev64ifbe(((int64_t*)is->contents)+pos);
     } else if (encoding == INTSET_ENC_INT32) {
-        ((int32_t*)is->contents)[pos] = value;
+        ((int32_t*)is->contents)[pos] = (int32_t)value;
         memrev32ifbe(((int32_t*)is->contents)+pos);
     } else {
-        ((int16_t*)is->contents)[pos] = value;
+        ((int16_t*)is->contents)[pos] = (int16_t)value;
         memrev16ifbe(((int16_t*)is->contents)+pos);
     }
 }
